Kodi Wink Integration
Dim the lights when Kodi starts playing. This flow will record what the
current status of the "Family Room" group is (on/off and brightness) and
will ensure that it only dims the lights if the lights are already on.
Furthermore, when Kodi stops playing media it will return your lights to
the brightness prior to watching the show.